+++ packages/docz-theme/src/components/Logo/styles.js
@@ -1,17 +1,18 @@
 export const logo = {
   bg: "blueDark",
+  color: "yellow",
   textAlign: "center",
   fontWeight: "bold",
   fontSize: 4,
   textTransform: "uppercase",
   py: 5,
   px: 7,
+  letterSpacing: 1,
 };
 
 export const link = {
-  background: "var(--color-yellow)",
-  WebkitBackgroundClip: "text",
-  WebkitTextFillColor: "transparent",
+  color: "yellow",
+  textDecoration: "none",
 };
 
 export const image = {
