+++ packages/docz-theme/src/components/Sidebar/Sidebar.jsx
@@ -1,6 +1,15 @@
 /** @jsx jsx */
 import { Box, jsx } from "theme-ui";
+import { Logo } from "@jobber/docz-theme/components/Logo";
 
-export const Sidebar = () => {
-  return <Box>Sidebar</Box>;
-};
+export function Sidebar() {
+  return (
+    <Box>
+      <Box>
+        <Logo />
+      </Box>
+      <Box>Search</Box>
+      <Box>Navigation</Box>
+    </Box>
+  );
+}
